New CNN and Improved threshold in classification
New CNN has same architecture but it is trained on MNIST with salt
and pepper, as well as randomly added horizontal and vertical lines.
Major improvement also by disabling reconstruction, which were causing
information loss. This was due to thresholding, which was removing the
digits when the value of the surrounding box was much higher.
See also note in relevant file.