fix visibility of grade comments
A previous commit introduced a bug on the `Old Comments` section of
the submission detail page. Only non-grade-summary comments are
visible in the submission detail page (since grade summary comments are
not very informative. But if a submission has only grade summary
comments, the title of the section (Old comments) was still visible,
despite the fact that there were no "regular" comments to display.
This commit fixes this issue.