Fix submission pdf attribute in new uploads
The files were stored at the intended file path, but the pdf property of
PaperSubmission that references the file should be the relative path
from media, not the absolute path of the pdf. That’s the reason why
if you went to the admin site, you can open and view the image of
PaperSubmissionImage, but you get an error message if you try to view
the pdf of PaperSubmission. Old pdf submissions remain unchanged
and their functionality is not affected. New pdf files keep the same
functionality but they are also renderable on the UI after this fix.
This bugfix will make PDF annotation support easier in the future.